Added some experiment fields to Material and Methods and remove some from Overview
